pyproject.toml
[project]
name = "ayol-uchun"
version = "0.1.0"
description = "qizlar-va-ayollar-uchun"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "amqp==5.3.1",
    "asgiref==3.8.1",
    "billiard==4.2.1",
    "black==25.1.0",
    "celery==5.5.3",
    "certifi==2025.7.14",
    "cfgv==3.4.0",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "distlib==0.3.9",
    "django==5.2.3",
    "django-environ==0.12.0",
    "django-filter==25.1",
    "django-jazzmin==3.0.1",
    "django-modeltranslation==0.19.16",
    "django-query-counter==0.4.1",
    "django-recaptcha==4.1.0",
    "django-rosetta==0.10.2",
    "djangorestframework==3.16.0",
    "djangorestframework-simplejwt==5.5.0",
    "drf-yasg==1.21.10",
    "filelock==3.18.0",
    "flower==2.0.1",
    "humanize==4.12.3",
    "identify==2.6.12",
    "idna==3.10",
    "inflection==0.5.1",
    "iniconfig==2.1.0",
    "kombu==5.5.4",
    "mypy-extensions==1.1.0",
    "nodeenv==1.9.1",
    "packaging==25.0",
    "pathspec==0.12.1",
    "pillow==11.3.0",
    "platformdirs==4.3.8",
    "pluggy==1.6.0",
    "polib==1.2.0",
    "pre-commit==4.2.0",
    "prometheus-client==0.22.1",
    "prompt-toolkit==3.0.51",
    "psycopg2-binary==2.9.10",
    "pygments==2.19.1",
    "pyjwt==2.9.0",
    "pytest==8.4.1",
    "pytest-django==4.11.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "pytz==2025.2",
    "pyyaml==6.0.2",
    "redis==6.2.0",
    "requests==2.32.4",
    "ruff==0.12.0",
    "six==1.17.0",
    "sqlparse==0.5.3",
    "tabulate==0.9.0",
    "tornado==6.5.1",
    "tzdata==2025.2",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
    "vine==5.1.0",
    "virtualenv==20.31.2",
    "wcwidth==0.2.13",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
    "ruff>=0.12.0",
]


[tool.ruff]
target-version = "py312"
line-length = 88
exclude = ["__pycache__", ".git", ".venv", "migrations", "build", "dist"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle (PEP8)
    "F",     # pyflakes
    "I",     # isort (import tartibi)
    "B",     # flake8-bugbear
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "C90",   # mccabe
]
ignore = [
    "E501",  # Juda uzun qatorlar uchun ogohlantirmasin
]

[tool.ruff.lint.isort]
known-first-party = ["myapp"]  # bu yerda o‘z loyihangiz modulini kiriting
force-sort-within-sections = true
combine-as-imports = true
